# bpo-37543: optimize pymalloc