PyObject_Malloc() and PyObject_Free() inlines pymalloc_alloc and
pymalloc_free partially.
But when PGO is not used, the compiler doesn't know where is the hot part
in pymalloc_alloc and pymalloc_free.

To help compilers:

* use LIKELY/UNLIKELY macros
* mark pymalloc_alloc and pymalloc_free as inline function
* split slow / rare parts into other functions

```
# bm_spectral_norm
Mean +- std dev: [master] 199 ms +- 1 ms -> [patched] 176 ms +- 1 ms: 1.13x faster (-11%)
```

<!-- issue-number: [bpo-37543](https://bugs.python.org/issue37543) -->
https://bugs.python.org/issue37543
<!-- /issue-number -->