@jdemeyer 2019-07-18
Did you putting the macros `LIKELY` and `UNLIKELY` in a public header file (renaming them of course)? There are surely other places in CPython that could benefit from them.